move util func to util.py
